//Core Elements: API Core parts, part two
This commit adds the following to part one:
 - Enhanced documentation for use with Sphinx (Napoleon)
 - Some small changes to the configuration system and related unit tests
